+++ include/misc/transform_vals.hpp
@@ -24,7 +24,7 @@ inline
 arma::vec
 transform(const arma::vec& vals_inp, const arma::uvec& bounds_type, const arma::vec& lower_bounds, const arma::vec& upper_bounds)
 {
-    const int n_vals = bounds_type.n_elem;
+    const size_t n_vals = bounds_type.n_elem;
 
     arma::vec vals_trans_out(n_vals);
 
@@ -56,7 +56,7 @@ inline
 arma::vec
 inv_transform(const arma::vec& vals_trans_inp, const arma::uvec& bounds_type, const arma::vec& lower_bounds, const arma::vec& upper_bounds)
 {
-    const int n_vals = bounds_type.n_elem;
+    const size_t n_vals = bounds_type.n_elem;
 
     arma::vec vals_out(n_vals);
